Updated openstack/openstack Project: openstack/oslo.db  de9f89b60c719698f4be36abe4cef299b96671b4
Adjust SAVEPOINT cause test for SQLA 1.1

SQLAlchemy 1.1 has repaired some of the failure modes
for Session.rollback(), including the MySQL case where the savepoint
identifier gets lost, such that the Session is in a more usable
state than before.  A side effect of this is that the test
here in oslo.db only hits a single error handler, not two.

SQLAlchemy upstream is described at:

https://docs.sqlalchemy.org/en/latest/changelog/migration_11.html#improved-session-state-when-a-savepoint-is-cancelled-by-the-database

Change-Id: I324f288701e9ffd241de638e0e0419dc8d9c1264