Updated openstack/openstack Project: openstack-infra/project-config  09d41349ccaa121f688402460cdc539a39e01384
make a non-voting gate-trove-functional-mysql-ubuntu-xenial job

The goal of this commit is to create a single non-voting
gate-trove-functional-mysql-ubuntu-xenial job that will only run in
the master branch of the trove project. It will be used to verify a
change that will make it possible to run the trove CI jobs for master
on xenial nodes.

Change-Id: Ibc0ac0ae205a748ffaa7fb4c8e4305d75515aae1