Updated openstack/openstack Project: openstack/neutron  d650b4a56ea2f9de843637d474957f149b693d8b
ovsfw: small cleanups to improve readability

security_group_updated is actually called but of no use for ovsfw.
Fix the comment to reflect the reality.  In iptables_firewall, the
method is used to clear obsolete conntrack entries, but ovsfw implements
that differently.  Also, remove empty definition of apply_port_filter,
which is never called.

Change-Id: I26b402e0fa5ac19022e022bf7c51e171d0f5d333
Closes-Bug: #1606462