Updated openstack/openstack Project: openstack-infra/shade  cb281e2a6c0c36209010bf31ce15016d362d3bba
Allow server to be snapshot to be name, id or dict

In documenting this method, it became clear we weren't telling people
what value server should have. Our tests made it look like it could be a
server name - but in fact it could actually only be a server dict or a
server id.

Make it explicitly all three, and fix the tests to not test their own
mocks. Or, if not actually fix them, at least make their examples not
blatantly wrong.

Change-Id: I64361a7a26cfa5137f9e862624fe379219f1cbb1