Updated openstack/openstack Project: openstack/ironic  2c4998e7b88fb74e407bce7cf8641c4737d53f20
Fix exception handling in iscsi_deploy.continue_deploy

When IronicException is raised, do not replace it with a more generic one.
When unexpected Exception is raised, log it as LOG.exception, not LOG.error.
Otherwise it may be impossible to figure out what went wrong, e.g.
in case of a stdlib exception.

Change-Id: I60ec259f56948e473c9b5c363372215edfe21dea
Related-Bug: #1626544