Updated openstack/openstack Project: openstack/nova  5da51f7dbbb008559a5ccdb14b5089715fdfce82
Use byte string or utf8 depending on python version for wsgi

A recent change to the wsgi code broke deployments running over
httpd/mod_wsgi. This is because for py2.X mod_wsgi accepts byte
strings and breaks with utf8. However, for py3.X, utf8 is
accepted. So this acts accordingly.

Change-Id: I81739bc3de9d623b718987b5fc18eaf851533902
Closes-Bug: #1643511