Updated openstack/openstack Project: openstack/manila  799fc8d55ffd409964ca8e2c956183c1b399f4d9
Fix extend operation of shrinked share in generic driver

If a share is shrank from 10G to 5G, the volume is still 10G.
This share`s size display is 5G, but the share extend newsize
must be > 10G in the old code. So, I think only need to perform
resize_filesystem when the share extend newsize <= volume size.
The volume extension is performed only when new size larger than
volume size.

Change-Id: I7e49b446445b8005e2eb23e1d439354eb24915e0
Close-Bug: #1639188