Updated openstack/openstack Project: openstack/tempest  ea35afb3e08bc6ce1dc6bafa6901b162f87b7049
Add missing tests for the image v2 API

This commit adds missing test cases for the image v2 API, the missing
test cases are for listing images using different parameters for
filtering the lists.

This commit also fixes one test which was attempting to validate the
use of the container_format filter using a value that was not found
in the test environment.

Change-Id: Ia1369fa73f56ec290e868d3f2f2873785a70951d
Closes-Bug: 1086590