Updated openstack/openstack Project: openstack/cinder  808862bf37612b1cbb45e649f1c7e551879b5c64
Add host check while create snapshot

We may create a volume but the volume was not been scheduled for some
reasons(e.g. The disk space is unavailable, cinder-volume service is
down, etc...). Although this volume is in error status, we can force
create snapshot using `--force` flag, that's make no sense. What's worse,
it may carry some unexpected error: We can delete this error volume before
cleanup the snapshot, but we can't delete the snapshot any more via API for
its prarent volume not found. This patch just raises an InvalidVolume
exception when we create a snapshot on a volume without host. In
addtion, the volume state is mutable so I think it is not 100% reliable,
the user may reset state w/o checking host state. I think we need check
host state before go forward to create snapshot on it.

Change-Id: Ifb9388cc76c1ff8d2e7f5b24710acbe95b602825
Closes-Bug: #1641294