Updated openstack/openstack Project: openstack/tempest  3f5d149b3a32e713c60c59a054035ac2e5c73c28
Merge the separated link lines on volume client

On current docstring of service clients, every link to API
reference is separated to 2 lines. That is unreadable on the
code and the documentation which is created from the docstring
also will be unreadablebecause of the separated link lines.

This patch merges the separated lines into a single line for the
readability.

Change-Id: I46c48d66c63ad8da1c5e9b245adbf2a6dd7ce9e1