Updated openstack/openstack Project: openstack/tripleo-heat-templates  22b01f75dd32ea910abba93ea5f62c81ae096996
Enables auto-detection for VIP interfaces

Previously the ctrl plane VIP would default to 'br-ex' which in non-vlan
deployments ends up being the wrong interface.  The public VIP interface
was also defaulted to 'br-ex' which would be incorrect for vlan based
deployments.  Since a user has already given the nic template (and in
most cases the subnet that corresponds to the nic) the installer should
be able to figure out which interface the public/control vip should be
on.

These changes enable that type of auto-detection, unless a user
explicitly overrides the heat parameters for ControlVirtualInterface and
PublicVirtualInterface.  Also, incorrect parameters from haproxy service
are removed.

Depends-On: I05105fce85be8ace986db351cdca2916f405ed04

Closes-Bug: 1606632

Change-Id: I3c1c39824ec32ced304a782edc6ef49c0769c108
Signed-off-by: Tim Rozet <trozet@redhat.com>