Updated openstack/openstack Project: openstack/neutron  908d81f076c97e0c3664c69cc28e4de0b38d642e
Fix security_groups_provider_updated for linuxbridge

According to [1], security_groups_provider_updated will get port ids
to update the related security groups.

However, in the firewall of l2 agent, ports are identified by 'device'
attribute. This is fine for ovs agent, because, for ovs agent, port id
is the same as 'device' attribute.

But for other agent, like linuxbridge agent, the 'device' attribute of
port will be the device name. This will make l2 agent can't find port
and fail to update the security group of port. The info message at [2]
can be observed.

The fix is to convert port id list to 'device' list.

[1] http://git.openstack.org/cgit/openstack/neutron/tree/neutron/db/securitygroups_rpc_base.py#n157
[2] http://git.openstack.org/cgit/openstack/neutron/tree/neutron/agent/linux/iptables_firewall.py#n177

Change-Id: Idf9bd7942ef2063cfbc5e9eeead8d49d09883ea8
Closes-Bug: #1555027