Updated openstack/openstack Project: openstack/swift  0873b7c03eddfa02c883a8b553abc6f2d5a26b1f
Document access control lists (ACLs)

Documents the syntax and meaning of container ACLs. Account ACLs
were already pretty well documented. However the account
ACL text was moved as part of this change.

TempAuth and keystoneauth have diffent ACLs. However, I decided
to describe both in one section/table because there are many
"examples" of ACLs in other documents, and it's better that
someone coming here from those sources become aware that the
specific ACL might not apply to them. In addition, the
referrer and .rlistings is common to both.

Some changes were also made to the api-ref document. The doc
and api-ref documents are published as seperate documents, so
the cross references from the api-ref section will not work
until this patch merges and the documents are rebuilt.

Change-Id: Icd2d6c278050c263b833ae76545c041f54fae68d