Updated openstack/openstack Project: openstack/nova  59642a5d76cb0753d34448a458cc7356eaaf81f4
Refactor two nearly useless secgroup tests

There are two tests that masquerade as "compute api tests" to
validate some instance security group associations. However, those
are just there because it's a convenient place to create instances
with a full complement of things. They depend on behaviors we're
about to move, and are also slated to be irrelevant when we remove
nova-network. In the meantime, move them closer to the other tests
for the bits they are exercising to get them out of the way of
the following patch.

Change-Id: I58f2d9d166446cd762a2fd4ad9d6b4bf5fff0c1d