Updated openstack/openstack Project: openstack/neutron  0c706d9eb6dd937c7d2f797e707868d4f464bb09
Separate floating IP port creation from transaction

This moves the floating IP port creation outside of the transaction
that creates the floating IP record. This eliminates the use of one
of the GUARD_TRANSACTION flags to prepare us for the enginefacade
switch and correct event notification semantics for push
notifications.

Note that this introduces a small window where the server can die
mid-process and leave an orphaned port without a floating IP. This
is similar to what can happen for router gateway ports. The intention
is to address these in a follow-up patch with a periodic garbage
collection task that will look for floating IP ports with a device
ID of 'PENDING' since the tenant cannot delete them on their own.

Related-Bug: #1540844
Partially-Implements: blueprint push-notifications
Partially-Implements: blueprint enginefacade-switch
Change-Id: Ia4c34c6654a5bfb64fbf06b11b0a29b018c6854f