Updated openstack/openstack Project: openstack/nova  194c77a42785b664f953bb5a352f8a423747e202
placement: adds ResourceClass.get_by_name()

The next patch in this series adds the ability to delete a custom
resource class. In order to implement that, we need a method to return a
single objects.ResourceClass object by string name. This patch adds that
method, implemented by using the resource class cache in a transparent
manner.

Change-Id: I92f84ed24518155a77f281057808750803c84889
blueprint: custom-resource-classes