Updated openstack/openstack Project: openstack/diskimage-builder  f15550f9fe2db8a9fe86fb89f28039aa5c46323b
Special case dib-python in dib-lint

It seems that on Xenial, it does not take much to confuse "file" and
it's mime guessing such that it thinks some files are not python.

"package-installs-v2" is a good example, since it has an interpreter
"dib-python" that "file" doesn't know about, and no extension.  While
looking at this, I've added emacs vars here so it opens in python
mode.

Change-Id: I01994b08c5ad8987925f1eec4062f5b6ee72eb8f