Updated openstack/openstack Project: openstack/neutron  f2235b7994b22d3e4be72185b86ba5723352f4b0
Move DVR fip agent gw port create out of transaction

This shouldn't be happening in a transaction. It creates
port resources in ML2 that need to have correct commit
semantics.

This refactors it to leverage the callback framework and
unifies the codepath for creating these DVR ports after
a router creation and after a router update with a migration.

Related-Bug: #1540844
Partially-Implements: blueprint push-notifications
Partially-Implements: blueprint enginefacade-switch
Change-Id: I15d8d32b54087e4a2ca1bfa74b46fa95adc4b95d