Updated openstack/openstack Project: openstack-infra/project-config  32d2e1658a44ab965122b0d2e685f5a51b091f1a
Fix different zmq configurations for functional tests

Currently the driver name affects the choice of the test
configuration script being used. As it comes from test results
the same configuration runs for zeromq, zeromq-pub-sub, zeromq-proxy
so it doesn't make sense as we would like to test different zmq
configurations on that gates.

Change-Id: Ibb9cdbba4b896c17703b249eb07ec8f5be583a85
Depends-On: I94342be65564665faf27d8d01ae6650183146874
Related-Bug: #1643929