Updated openstack/openstack Project: openstack/tripleo-puppet-elements  cb1706efcf6fc9c4713d40e722d8a55a4ecc7f35
Element overcloud-agent replaces os-collect-config

The element overcloud-agent is a purely package driven replacement for
the tripleo-image-elements element os-collect-config.

All it does is install the python-heat-agent package and enable the
os-collect-config service.

The intention is for this element to be merged unused, then a change
to the tripleo-common/image-yaml will confirm that it is working as
expected.

This is part of the effort to deprecate the entire
tripleo-image-elements repo.

Change-Id: I1f3bb386660f36813515544dc73b1c5e08467553