Updated openstack/openstack Project: openstack/tripleo-heat-templates  145961271715bc16e65072a5092ddba54694efa5
Configure Keystone Fernet Keys

Provision the Keystone Fernet Token provider
by installing 2 keys with dynamic content
generated by python-tripleoclient.

Note that this only sets up the necessary keys to use fernet as a token
provider, however, this does not intend to set it up as the default
provider; This will be discussed and will come as part of another
commit.

Co-Authored-By: Juan Antonio Osorio Robles <jaosorior@redhat.com>
Depends-On: Ic070d160b519b8637997dbde165dbf15275e0dfe
Change-Id: Iaa5499614417000c1b9ba42a776a50cb22c1bb30