Updated openstack/openstack Project: openstack/neutron  d343354f7ff94a77d62f9cb0cb593e7187122085
Get post-subnet actions out of transaction

ML2 was enclosing the call to super().create_subnet in a
transaction in order to call its own extension processing
logic and mechanism driver precommit logic in the same transaction.
This had the side effect of putting things in a transaction that
the DB base plugin expected to call after the subnet was created.
These are update_port calls and router gateway update calls.

To get the correct transaction semantics and address a standing
TODO in the codebase, this adds _create_subnet_precommit and
_create_subnet_postcommit methods to allow the core plugin inheriting
from db_base_plugin_v2 to control the timing of the calls to the
postcommit actions.

This will also unblock some work for the new engine facade work and
the push notifications work.

Partially-Implements: blueprint push-notifications
Change-Id: I28323eddcf34fd12002cc706f4a9bcbb3f976ceb