Updated openstack/openstack Project: openstack-infra/irc-meetings  ba10ad453dbd93b8a734ac025734365bd291c45c
Remove third-party-ci working group meeting

There are two third-party CI oriented meetings: this one and the one
on Mondays 15:00. The Monday meeting became more of an office hours,
while this meeting's slot was used for discussing any work we can do
as third-party CI operators to make our life easier.

There has been not much attendance of either. So we remove this meeting
in favor of using Monday slot when necessary.

More details are in the ml announcement:
http://lists.openstack.org/pipermail/openstack-dev/2016-November/107859.html

Change-Id: I70a5e7a0f5267d2a4e69d5a10ffb00d8c95ebc6d