Updated openstack/openstack Project: openstack-infra/puppet-zanata  c5bddd9e193806a40cfe12b3abed2704abb88f7a
Hack to support service on xenial

Xenial uses systemd which does have a sysv init script compatbility
layer but you must instruct systemd to reload its config after
installing any init scripts for this to work. Hack this in by adding an
exec for `systemctl daemon-reload` between writing the file and starting
the wildlfy service. Ideally the wildfly module would do this for us but
this should work for now (and eventually will become redundant).

Change-Id: If5c09c0be7714ce7d03ebf364823164a34c29ae4