Updated openstack/openstack Project: openstack/nova  f7bee5ff2f978998f5c07cc5a31b1f0b61fbb6c5
Add {get_,set_}aggregates to objects.ResourceProvider

The Placement API will provide support for associating a list of
aggregate uuids with a resource provider. This is a way of stating
that, for example, a shared storage provider can be used by
compute hosts in aggregates A, B and C.

This change implements the basic get and set functionality required
on the ResourceProvider object as callable methods. Using fields was
also considered but given the precedent with set_inventory and
friends, methods were chosen. Because methods were chosen, explict
get and set-based names were used to avoid confusion with fields or
properties. Since the methods will only be called from the placement
API service, they are not remotable.

The aggregates are stored in the PlacementAggregates model, which is
a recently introduced new table (in Change
Id0355cb022f68e962af306ff04cf724d22b68d19).

As the guts of the complexity are in the database handling, tests
are on the functional side.

Change-Id: I95258e632f2c034e5b7b803a753f393ad715ca50
Partially-Implements: blueprint generic-resource-pools-ocata