Updated openstack/openstack Project: openstack-infra/storyboard-webclient  f9b68493167b856644a9ed90d58562a29c3e0bed
Make subheading more succint

We wanted this to be more detailed, to make it clear to users what
'task-tracking' referred to. That's not really what a title is for, though,
and there's not enough room to do it justice in the subheading. This patch
has a subheading that assumes familiarity and gets to the point; if we need
a more thorough explanation, we should link to a detailed one elsewhere.

Change-Id: I89c087e5a1b6d90cd7cad1646a1f0ae4a7d4a7b6