Updated openstack/openstack Project: openstack/nova  ce23061488a8535b211bc86cdd5d2feaadfc75be
Implement get and set aggregates in the placement API

/resource_providers/{uuid}/aggregates
GET for a list of aggregate uuids associated with this resource
    provider
PUT to set the list of aggregate uuids associated with this this
    provider

The API requests map directly to set_ and get_ aggregates on the
ResourceProvider object.

This is implemented as placement API microversion 1.1. To make that
easier a raise_404_if_not_version helper method is added to the
microversion module.

The new rest_api_version_history doc is updated to reflect this
new version.

Change-Id: I514c15c7d387cf25bf7986d07baccf0e7a785f46
Partially-Implements: blueprint generic-resource-pools-ocata