Updated openstack/openstack Project: openstack/qa-specs  af2786f3159f76b999ed28cf5cf5ae7dbb732ff2
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

b'https://gist.github.com/cff28059637f37993802631cb30976d1\n'

Change-Id: I024f7fb9f72bf05dacbf9b43323eb3c29206894a