Updated openstack/openstack Project: openstack/ironic  50726dbcb00ff69a4409c552afc351e5f66b3d34
Minor changes to neutron security groups code

This is a follow-on patch to 3197e44c04de064bc3d8af09af7e0d2d9511af6d.
It cleans up a bit of the code and addresses the nits (changes a
LOG.exception to LOG.error and adds a unit test).

Change-Id: I02b6346d9a2abff858c9dd6083fd29f393c63e97
Partial-bug: #1594242