Updated openstack/openstack Project: openstack/heat  19133101e16111d8fc0c99646d316cc77bcd9374
Use enginefacade instead of db_session

This change takes a gradual approach to fixing bug #1479723 by
creating an enginefacade transaction_context, but mostly still using
the legacy facade for getting the default session.

Where a separate independent transaction is needed,
db_context.writer.independent.using is used instead of get_session.

The next change will decorate the db api functions with reader and
writer so queries in a function are performed in a single transaction.

Change-Id: If92788641d84b7314ed7ad1f8cb0e141fa1a8681
Partial-Bug: #1479723