Updated openstack/openstack Project: openstack/tripleo-heat-templates  a864074f28a9ff802989f1d18387e927c265e2d6
Make Ceilometer notifications non-blocking

Ceilometer notifications can be sent in a background thread, unblocking
the Swift proxy in case the RabbitMQ is not processing notifications
quick enough or even unavailable.

There is a default queue size of 1000 notifications. If more messages
are added to the queue these will be discarded, and a warning log entry
will be emitted.

Change-Id: I98022dcbf661a5bb7425f49ba8525225d61212dc