Updated openstack/openstack Project: openstack/nova  a522458ba0a9d46237d752bd13073cb5046e1276
libvirt: handle os-brick InvalidConnectorProtocol on init

On init the LibvirtDriver imports it's volume drivers into
a registry for later access during volume operations. Several
of the libvirt volume drivers are using os-brick for the connector.
os-brick can raise an exception when loading a connector class
under certain conditions, e.g. s390x + ISER is not a supported
mapping in os-brick. This in turn causes the LibvirtDriver to
fail to load which causes nova-compute to crash.

This change handles the os-brick specific InvalidConnectorProtocol
exception when importing volume drivers, logs a debug message but
doesn't fail to init the LibvirtDriver.

As part of this change the driver_dict_from_config helper method
is integrated with the _get_volume_drivers method as that was
the only thing using it, and it's cleaner if those are just a
single method.

Depends-On: I7183a479bc12c4e5509f52bf7fdd0fa7ccaa7247

Change-Id: I61376469a409dc5475886e6c79486b815129392a
Closes-Bug: #1639239