Updated openstack/openstack Project: openstack/nova  eb4f9a44c79494d8df936511380f351eb3a0301e
libvirt: Delete the lase_device of find_disk_dev_for_disk_bus

Delete the last_device parameters of find_disk_dev_for_disk_bus.
Just pick the first available letter on the bus as the device,
instead of hardcoding on the last device for config drive. Also
tweak the test case for this.

Closes-bug: #1591827

Change-Id: I6343d7ab6f1b0aa58f29805e24db7e2c61d1409e
Signed-off-by: Kevin Zhao <kevin.zhao@linaro.org>