Updated openstack/openstack Project: openstack-infra/project-config  8bd15729b577e022033d990ca50ea66949d2b4e3
Create repo for Interop Challenge assets

With the success of the Interop Challenge in Barcelona, the Interop
Challenge mission continues to gain momentum.  We would like to have our
own repo where we can focus on creating and iterating on new
OpenStack workloads that will be used to demonstrate that OpenStack
clouds are interoperable across a variety of application domains
including enterprise workloads, NVF workloads, K8s, and many others
under evaluation. We will also use this repo to store our best practices
documents and other collateral that facilitates improving the state of
OpenStack Clouds' ability to interoperate.

Change-Id: I93c3603ab6c0ea486fc9af53fab5fdc3a1a340db