Updated openstack/openstack Project: openstack/requirements  09519110784ea21ecf71f4085fe8ad7b17a4fedf
Bump the minimum requirements for pip and virtualenv

As a whole OpenStack requires constraints support to do any CI or
development.  Our current specifications do not enforce that.

constraints support was released in pip 7.1.0, and pip 7.1.0 was
bundled/vendored in the 13.1.0 release of virtualenv.  You can see from
the output below that from a virtualenv POV we meet the minimum on all
branches

 ---
 Requirements
 origin/master         : virtualenv  # MIT
 origin/stable/newton  : virtualenv  # MIT
 origin/stable/mitaka  : virtualenv  # MIT
 origin/stable/liberty : virtualenv

 Constraints
 origin/master         : virtualenv===15.1.0
 origin/stable/newton  : virtualenv===15.0.3
 origin/stable/mitaka  : virtualenv===14.0.6
 origin/stable/liberty : virtualenv===13.1.2
 ---

Change-Id: I45eee626713438af5fc676f2b5f636d7ec23f7be