Updated openstack/openstack Project: openstack/python-designateclient  0a0e9725b1dbfb521cf28c1efc379df30df90102
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/11fc0b132d352e49f6aa4e375d2a5415

Change-Id: I2ce3317d48fcd4e3755b57dbb524ec715d881a1f