Updated openstack/openstack Project: openstack/neutron-specs  63861abe7b2d7d1290c5ecf1b542a015e32686a2
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/146757313f829f0af165d6bf93ffd428

Change-Id: I0ac9c536c1b8f5558ed35fc4c778bf03f8cdf04b