Updated openstack/openstack Project: openstack/neutron-lbaas  da625cd6463dbbb1b6047de0c746936c20fd9d17
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/9eb1f97639a2df787abcd5573f56a253

Change-Id: I03f0bad2ed596d238ebc1cc84455c7ceb6f0e688