Updated openstack/openstack Project: openstack/swift-specs  e53112c2718b8a5db5571867b01e51acff42dd9f
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/b1e11e7cb7c0d9aaecbfc2f46859a8c8

Change-Id: I48f32ced21f66c95213858cbbaaf96cff04f0bf5