Updated openstack/openstack Project: openstack/python-keystoneclient  660e02f25972ee4baa974e6efa2c5f6bc9cf68be
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/9ccb4a8d24c50452d5c816c357a4c220

Change-Id: I79ea8e8aafe99d729b7ee64899b6e7342afc40ee