Updated openstack/openstack Project: openstack/pycadf  2e89960017cc6715abb8132aab86da8568cda2bc
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/f7a020d80d955f21b95d43909b57ae0d

Change-Id: I5207741cee7f50e817feedbb8dad252e742400f3