Updated openstack/openstack Project: openstack/keystonemiddleware  4ffbc2224187fa17dedd9b5024ec36b3903249fc
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/23379e562e65f9be0f8853f40d2e61b3

Change-Id: I6a1518a7a5ad3d05458aa5b9d23025924a7ea965