Updated openstack/openstack Project: openstack/keystone-specs  076be6ffd69ea8d21b53c6d1b9eec47df9a00281
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/7c2de9e833cbaf2a6211bb42f1928654

Change-Id: I0517744d69edad0c2f80f3102353f91ff8e26a47