Updated openstack/openstack Project: openstack/heat  c8cea960895a9e4789927b7c2dbe3c0688b8b0ad
Don't use oslo-incubator in glance client plugin

There is plan to remove all oslo incubator code from the
clients. This removes the usage of deprecated
openstack.common.apiclient.exceptions in glance client
plugin.

The available glance HTTPNotFound exception is not usable
for validations(no message) and there is no UniqueMatch
exception.

This adds a new custom client_exception module with two
exceptions to ensure that the error messages remain the
same as before.

Change-Id: I5b32eef047a1fe38382f649532c8bb8a910315e5