Updated openstack/openstack Project: openstack/cliff  25a4589d36bbdd769e6ca9907eb3acff6b5eef1e
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/f4a7414b5a764600f25f7c75cec24919

Change-Id: If1b8c1472323dd08d5b3eff0335b7a32a42e2a18