Updated openstack/openstack Project: openstack/heat-cfntools  afaf465ac5d9d9f05aa57a0157d889f73917c864
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/d71e2eaf72170035e6d10135f9c04a02

Change-Id: Icb4f114e9408cd84e885aa51c3f1a9bfb54d7c52