Updated openstack/openstack Project: openstack/sahara-dashboard  3d7ed540c567eef7da0b3c66b1e2eae16a915ca5
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/d4c8e744698ee0ef79d02b5390f7d707

Change-Id: Ibd183d1e5cb2734c3b9b1d17b198d875c9664fa4