Updated openstack/openstack Project: openstack/sahara-extra  ac8b0480df83673a9d0018b0af13dcdad8f94d41
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/b8a9835f9193c610b76acdf617aecc40

Change-Id: I4202a0e91d951debdf33ef5224d98bab21121aa1