Updated openstack/openstack Project: openstack/python-saharaclient  f624de6a1e30dfb5a6f3a633684d78079362dd0a
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/26cc2ee8b0a8e4a96cd0c63ac768c0e1

Change-Id: Iee796b3b6013b2dee84fb63f8131bc50f0d1faa6