Updated openstack/openstack Project: openstack/cinder  514aff0f5e47dcf48ea98a692d83f8fb4cf4e266
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

b'https://gist.github.com/67f4c17475042794640f06833d929bba\n'

Change-Id: Ib733302fc5373618a778f22add9512c36b1a6de9