Updated openstack/openstack Project: openstack-infra/project-config  40486e37f3015450d3f84bb6e8ed7daefceeb972
rename gate-trove-scenario-functional jobs

scenario jobs are by definition functional and the names are getting
long and messy. simplifying gate-trove-scenario-functional- and making
them gate-trove-scenario- instead so when I add the node prefix change
(coming to a theatre near you very soon) the names are not too long.

ditto for the legacy jobs.

Change-Id: I149434240fdf4c0c4110186e7bc86d38fcac3895