Updated openstack/openstack Project: openstack-infra/project-config  331ca88c06198601116f62d359c8a19254b6bba0
Have scenario tests run in two steps

The scenario tests currently run in a single stream and as we add more
tests, this takes a long time to complete. Changes in the way the jobs
are defined in [1] make it possible for us to split tests into
subgroups and run them independently and this is important as we
continue to add more scenario tests.

This change makes the CI jobs leverage that capability. The changes
that make this possible (in trove) were committed in [1] and a
dependency is therefore added here.

As we are midway in the trusty->xenial transition, this commit is
stacked on that pile; when the transition to xenial is complete, the
trusty jobs here will be removed and the xenial jobs made voting.

[1] https://review.openstack.org/#/c/399773/

Change-Id: Iba32f34b2ad70c38951668372a2e04eefb193720
Depends-On: Iea2a996043b4c2d1889995e160891af33ffab5b6
Co-Authored-By: Petr Malik <pmalik@tesora.com>
Co-Authored-By: Peter Stachowski <peter@tesora.com>
Co-Authored-By: Amrith Kumar <amrith@tesora.com>