Updated openstack/openstack Project: openstack/trove-integration  bd3aafabd5b27c36ef8c9cc4b2a70dcc524eb5e7
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/af3bcb242c9e2828ab264ec09beb108e

Change-Id: Ic30c28d25d8fa5fca20fdda8dedf1f92a1e6e2a2
Partial-Bug: 1644851