Updated openstack/openstack Project: openstack/dib-utils  694d2988cb62e1fc2d10631b9dea6ecba6c53991
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/1639ffc79be0cebf7ba162468f6c9ee5

Change-Id: I38d14e650166fbb4e5d8d4bcf8c9f331e7255937