Updated openstack/openstack Project: openstack/tripleo-image-elements  c538d271fabfc23a45d90b07eac2b1be1e94bf08
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/02c6146aafbcd4ec707e7b5d0c225a8c

Change-Id: I1a5c743afa0edd07caba377564e5d201694aeecb