Updated openstack/openstack Project: openstack/sahara  e146093b23ff17ab13520ed1bfac1da2465103c5
Show team and repo badges on README

This patch adds the team's and repository's badges to the README file.
The motivation behind this is to communicate the project status and
features at first glance.

For more information about this effort, please read this email thread:

http://lists.openstack.org/pipermail/openstack-dev/2016-October/105562.html

To see an example of how this would look like check:

https://gist.github.com/6f5a28b7577f9aea97e66d42c6df1e39

Change-Id: I1ff5f6ab38e0645c9105ecf4369648eb3c853f82